+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 Tracks changes of the codebase. 
 
+### 0.5.0 - (2026-01-28)
+
+* Expands `llama.cpp` benchmarks to include two MoE models.
+* Adds `ik_llama` benchmarks, a more performant fork of `llama.cpp` that reportedly does well
+on Pascal GPUs.
+* Adds gdsio benchmarking to measure how quickly data can be loaded from disk into VRAM.
+
 ### 0.4.0 - (2026-01-25)
 
 * Adds `compare` mode for visualizing difference between runs.